requirements.txt
accelerate
datasets
diffusers>=0.26.3
fire
gradio
jaxtyping
numba
numpy
git+https://github.com/NVlabs/nvdiffrast.git
omegaconf>=2.3.0
onnxruntime_gpu
opencv_python
opencv_python_headless
ort_nightly_gpu
peft
Pillow
torch==2.1.0 --index-url https://download.pytorch.org/whl/cu118
torch_scatter --index-url https://data.pyg.org/whl/torch-2.1.0+cu118.html
pygltflib
pymeshlab>=2023.12
git+https://github.com/facebookresearch/pytorch3d.git@stable
rembg
tqdm
transformers
trimesh
typeguard
wandb
xformers
